Question1/requirements.txt
streamlit
torch
transformers
evaluate
nltk
requests
cohere
google-generativeai
groq